Remove deadcode in zipkin translation (#1533)
